fix di alcuni problemi legati a livewire e non funziona il salvataggio in DB->da fixare

+++ database/migrations/2025_07_18_184319_create_articles_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('title');
             $table->text('description');
-            $table->float('price',8,2);
+            $table->decimal('price', 8, 2);
             //relazione con category
             $table->unsignedBigInteger('category_id')->nullable();
             $table->foreign('category_id')->references('id')->on('categories');
